auth_api/python/requirements.txt
Flask
pyjwt
pymysql
sqlalchemy
tomli ; python_version < "3.11"
gunicorn
